throw unsupported when trying to create or delete a bucket

+++ plugins/nf-amazon/src/test/nextflow/cloud/aws/nio/AwsS3NioTest.groovy
@@ -352,8 +352,7 @@ class AwsS3NioTest extends Specification implements AwsS3BaseSpec {
         if( target ) Files.deleteIfExists(target)
     }
 
-    @Ignore //FIXME
-    def 'should create a directory' () {
+    def 'should throw unsupported when create directory is a bucket' () {
 
         given:
         def bucketName = getRndBucketName()
@@ -362,10 +361,12 @@ class AwsS3NioTest extends Specification implements AwsS3BaseSpec {
         when:
         Files.createDirectory(dir)
         then:
-        existsPath(dir)
+        thrown(UnsupportedOperationException)
 
         cleanup:
-        deleteBucket(bucketName)
+        if (existsPath(dir)) {
+            deleteBucket(bucketName)
+        }
     }
 
     def 'should create a directory tree' () {
@@ -438,15 +439,14 @@ class AwsS3NioTest extends Specification implements AwsS3BaseSpec {
         deleteBucket(bucketName)
     }
 
-    @Ignore // FIXME
-    def 'should delete a bucket' () {
+    def 'should throw unsupported when trying delete a bucket' () {
         given:
         final bucketName = createBucket()
 
         when:
         Files.delete(s3path("s3://$bucketName"))
         then:
-        !existsPath(bucketName)
+        thrown(UnsupportedOperationException)
 
     }
 
@@ -472,7 +472,6 @@ class AwsS3NioTest extends Specification implements AwsS3BaseSpec {
         deleteBucket(bucketName)
     }
 
-    //@Ignore // FIXME
     def 'should throw a NoSuchFileException when deleting an object not existing' () {
 
         given:
